Prix journalier voiture en centime pour éviter les arrondis de calculs

See merge request but22/s4/s4-a-01/location-de-voitures!49

+++ server/database/migrations/2025_06_05_122951_create_car_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->decimal('car_mileage', 10, 3);
             $table->string('car_picture')->nullable();
             $table->boolean('car_default')->nullable();
-            $table->decimal('car_price', 8, 2);
+            $table->unsignedBigInteger('car_price');
             $table->enum('car_disponibility', [Disponibility::DISPONIBLE->value, Disponibility::EN_REPARATION->value, Disponibility::EN_MAINTENANCE->value, Disponibility::INDISPONIBLE->value]);
             $table->unsignedBigInteger('agency_id');
             $table->foreign('agency_id')->references('id')->on('agency')->onDelete('cascade');
